implement _call method to simplify calls to requests

+++ kitovu/api.py
@@ -11,39 +11,37 @@ class Api():
         self.headers = {'Accept': 'application/vnd.github.v3+json',
                         'Authorization': 'token {}'.format(self.token)}
 
-    def _get(self, url, **kwargs):
-        r = requests.get(url, headers=self.headers, **kwargs)
-        if not r.ok:
-            raise ApiError(r.text)
-        return r
-
-    def _put(self, url, **kwargs):
-        r = requests.put(url, headers=self.headers, **kwargs)
-        if not r.ok:
-            raise ApiError(r.text)
-        return r
-
-    def _delete(self, url, **kwargs):
-        r = requests.delete(url, headers=self.headers, **kwargs)
+    def _call(self, verb, url, **kwargs):
+        if verb == 'get':
+            r = requests.get(url, headers=self.headers, **kwargs)
+        elif verb == 'put':
+            r = requests.put(url, headers=self.headers, **kwargs)
+        elif verb == 'post':
+            r = requests.post(url, headers=self.headers, **kwargs)
+        elif verb == 'delete':
+            r = requests.delete(url, headers=self.headers, **kwargs)
+        else:
+            msg = 'the verb {} has not yet been implemented'
+            raise SystemExit(msg.format(verb))
         if not r.ok:
             raise ApiError(r.text)
         return r
 
     def get(self, uri, **kwargs):
         endpoint = self.hub + uri
-        r = self._get(endpoint, **kwargs)
+        r = self._call('get', endpoint, **kwargs)
         while True:
             yield r
             paging = Paging(r.headers.get('link'))
             if paging.next_link:
-                r = self._get(paging.next_link)
+                r = self._call('get', paging.next_link)
             else:
                 break
 
     def put(self, uri, **kwargs):
         endpoint = self.hub + uri
-        return self._put(endpoint, **kwargs)
+        return self._call('put', endpoint, **kwargs)
 
     def delete(self, uri, **kwargs):
         endpoint = self.hub + uri
-        return self._delete(endpoint, **kwargs)
+        return self._call('delete', endpoint, **kwargs)
